drf monitor

+++ pkg/custom/monitor/tenantMonitor.go
@@ -161,7 +161,7 @@ func (m *FairnessMonitor) Save() {
 			cellName = fmt.Sprintf("%s%d", idLetter, placeNum)
 			if masterResource, existed := events.MasterResourceAtTime(timestamp); existed {
 				currentMasterResource[username] = masterResource
-				log.Logger().Info("add master resource in tenants trace", zap.String("tenant", username), zap.Uint64("new master resource", masterResource))
+				log.Logger().Info("add master resource in tenants trace", zap.String("tenant", username), zap.Float64("new master resource", masterResource))
 			}
 			masterResource := currentMasterResource[username]
 			log.Logger().Info("master resource of specific timestamp", zap.Uint64("timestamp", timestamp), zap.String("tenant", username), zap.String("cellName", cellName), zap.Uint64("master resource", masterResource))
